backend/requirements.txt

fastapi
uvicorn[standard]
google-adk
python-jose[cryptography]
bcrypt
python-multipart
pydantic
pydantic-settings
python-dotenv
google-genai
supabase
motor
pymongo[srv]
certifi
